requirements.txt
Flask~=3.0.3
Flask-Bootstrap~=3.3.7.1
Flask-Moment~=1.0.6
Flask-WTF~=1.2.1
python-dotenv~=1.0.1
